remove dup code
