+++ src/utils.js
@@ -1258,3 +1258,34 @@ export function setOnAny(collection, key) {
   }
   return undefined;
 }
+
+/**
+ * Returns the type of the connection.
+ * @returns {string}
+ */
+function getConnectionType() {
+  const connection = navigator.connection || navigator.webkitConnection;
+  if (!connection) {
+    return 0;
+  }
+  switch (connection.type) {
+    case 'ethernet':
+      return 1;
+    case 'wifi':
+      return 2;
+    case 'cellular':
+      switch (connection.effectiveType) {
+        case 'slow-2g':
+        case '2g':
+          return 4;
+        case '3g':
+          return 5;
+        case '4g':
+          return 6;
+        default:
+          return 3;
+      }
+    default:
+      return 0;
+  }
+}
